A 16-bit wrap the C lost, and the machine stopped dead
Boran's level 12 solution ran at fourteen page flips in twenty-five seconds
where the editor does seven hundred. Not a crash and not a trap - the port just
stopped getting anywhere, which is the hardest shape to go looking for.

gdb on the hung process named it in one line: `start_sequence` at 0x26783, in
the loop that shuffles the sequencer's table down to make room for a new entry.

    for (si = 0x38; si + 4 != di; si -= 4)

The original computes that comparison in a register - `mov bx,si / add bx,4 /
cmp bx,di` at 0x269ee - so it is sixteen bits and it wraps. With `di` zero the
walk runs down past zero, BX comes out 0, and it stops after fifteen turns.
Written as `si + 4` in C the addition promotes to `int`, 0xfffc + 4 is 0x10000
rather than 0, and the loop never ends: it ran off the bottom of the table
writing pairs of words over guest memory until the frame rate collapsed.

`di` is zero when the new sequence's key is at or below the first entry's,
which is what `play_sound(12)` out of `part_step_1e5c` does on that level. So
it needed a particular sound from a particular part, and no screen comparison
would ever have gone near it.

Fixed by taking the comparison back to sixteen bits, which is what the original
does. All twelve of Boran's solution snapshots now run their machines, solve
their puzzles and take ADVANCE through to the next briefing - level 12 ends on
puzzle 13 with a score of 466.

CLAUDE.md already says the widths are the one fact the types carry and that
getting them wrong compiles, runs, and loses it silently. This is that, in a
comparison rather than a variable: `si` was the right type and `si + 4` was
not.

I looked for others of the same shape and found none.

+++ reconstruct/seg2619.c
@@ -395,7 +395,21 @@ void start_sequence(uint16_t es, uint16_t ax, uint16_t cx)
             break;
         other = FAR_PTR((uint16_t)SND16(0xa + di), (uint16_t)SND16(8 + di));
         if (other[0x15c] <= key) {
-            for (si = 0x38; si + 4 != di; si -= 4) {
+            /*
+             * **The comparison is 16 bits and has to wrap.** The original
+             * computes it in BX - `mov bx,si / add bx,4 / cmp bx,di` at
+             * 0x269ee - so when `di` is 0 the walk runs down to 0xfffc, BX
+             * comes out 0, and it stops. Written as `si + 4` in C the addition
+             * promotes to `int`, 0xfffc + 4 is 0x10000 rather than 0, and the
+             * loop never ends: it ran off the bottom of the table writing
+             * pairs of words over guest memory until the frame rate collapsed
+             * from thirty a second to one every two seconds.
+             *
+             * Reached by a sound played from a part's step - `play_sound(12)`
+             * out of `part_step_1e5c` - with the new sequence's key at or
+             * below the first entry's, which is what makes `di` zero.
+             */
+            for (si = 0x38; (uint16_t)(si + 4) != di; si -= 4) {
                 SND16(si + 0xc) = SND16(si + 8);
                 SND16(si + 0xe) = SND16(si + 0xa);
             }
